chore: extract shared fixtures from two oversized indexer test files
- What: the setup shared by every describe block moves out of
  `worker.test.ts` (265 → 206) and `api.test.ts` (269 → 226) into
  `worker-harness.ts` and `api-harness.ts`. Both allow-list entries removed.

- Why: #2833, and the project's own test convention — CLAUDE.md asks for
  "nested, one behavior per file", which a 269-line file with five describe
  blocks and 45 lines of preamble is not.

- Deliberately NOT named `*.test.ts`: bun collects by filename, so a harness
  named `*.test.ts` would be run as a suite with no assertions. Verified —
  `bun test src/indexer/__tests__/worker-harness.ts` reports
  "Tests need .test, _test_, .spec or _spec_ in the filename".

- Impact: none. Pure move; the same fixtures, imported instead of inlined.

Gate: `bunx tsc --noEmit` exit 0 · `bun test --isolate tests/build/ src/indexer/
tests/indexer/` 171 pass.

Allow-list: 22 at #2833's filing → **17**.

Refs #2833
